+++ main.h
@@ -8,7 +8,7 @@ class CClass
 {
 	public:
 
-	EXPORT void * operator new(unsigned int size);
+	EXPORT void * operator new(size_t size);
 	EXPORT void operator delete(void*);
 	EXPORT virtual ~CClass();
 };
